Add ability to use HTTPS port
As long as we have HTTPS enabled by default on master node,
it would be nice to use secure channel for nailgun agent. So,
this commit switch to HTTPS as default protocol for agent and
allow to use plain HTTP as a fallback.
We use SSL no_verify for connection cause now certificate for
master node nginx generated on container creation and we don't
have an ability to easily push it to trusted on bootstrap node.

Change-Id: I570890b19eac74db6f3666af966e86f64327740d
Closes-Bug: #1480866