Reboot if master node was reinstalled
If master node is reinstalled, the bootstraped node
has outdated MCollective config and as a result it
can not be used in any deployment, but is registered as
available in Fuel. This change fixes that by making
the agent monitor master node UID and reboot itself if
the UID has changed.

Change-Id: Iacb5eca57e339f973f676712766cb613a80f393e
Closes-Bug: #1431386