Accept removable block devices by PCI vendor ID
The issue is that, for particular raid controller obtaning vendor of
block device from sysfs attributes is not realible.

Once every raid controller is connected to PCI bus directly,
it's more realible to look at vendor ID of that controller to which
removable block device belongs.

For example, Adaptec is a perfect canditate to be added to a list of
removable PCI vendor IDs, since it's manufacturing RAID controllers
mostly.

Change-Id: Ib1018abbb46c65d909e594ad1d8921c8c5a139fc
Related-Bug: #1486601