Exclude USB block devices by the default
All USB storage devices must be filtered by the default as often this
type of devices can be just an emulated temprorary storage for FW
upgrade and so on.

If one wants to get usb block devices reported to nailgun, then
it could be either a cmdline option report_usb_block_devices or
the same option added to the agent' config file.

DocImpact
Change-Id: Id609715732fd0ab393d1557b4810464fbfaf096e
Closes-Bug: #1543221