Add --dry-run option to nailgun-agent
It's a bit hard to debug nailgun-agent, because it
tries to send info over http somewhere and sleeps at
the beginning. So you can't run it on your desktop, for
example.

Let's add --dry-run option, to have ability to just get
and print all information.

This is a small improvement, so no bug or blueprint.

Change-Id: If7309635d40ff3263a671fddd7df20efd917097c