Fix post-deployment pxe interface detection
W/o this fix node do not respond with proper PUT for pxe interface
as the pxe interface gets plugged into br-fw-admin bridge and
it is not reported as pxe-related one.

We should check if the interface is a part of bond and/or bridge
configuration and respond correctly.

Change-Id: Ifc1c396b0945fc5a42165b969b6924dcff5975b2
Closes-bug: #1581517
(cherry picked from commit 5fa721206baf3fad3fec424b0cdbe5013238c195)