Add 'Container' field for MD parser
The logic of filtering of fake RAID MD heavily relies on the
presence of of 'Container' field inside of parsed data.
If this field is missing, it will never figure out the name of any
fake RAID devices and its component.

This patch adds this field to a parser.

In addition to that, it also logs all found fake RAIDs and
their components for the sake of easy debugging.

Change-Id: I2066c5a0e995e542271cd308c9d83e2373787be4
Closes-Bug: #1617071
(cherry picked from commit 60173537665464db260fb93829eb8affced33535)