WA for udevadm settle loops
There is an issue with 'udevadm settle' command described at [1]. It
turns out that inconsistent task IDs stored in queue.bin and
uevent_seqnum could cause udevadm settle to loop trying to find
the task that doesn't exist. Kernel developers are skeptical about
this issue and not willing to fix it, but even if they will, we
still will take it to MOS.

By default, it uses a timeout of 180 seconds that is inconsistent
with a set of existing nailgun-agent's timeouts. We can change this
timeout and I think that 15 sec will be a reasonable value

[1] https://lists.gt.net/linux/kernel/1524376

Change-Id: I0e71786c3516463496a313ea17438582a45ad2aa
Closes-bug: #1665584